requirements.txt
azure-functions
pandas==2.2.3
openpyxl==3.1.5
python-dateutil==2.9.0
pytest==8.3.4
